ai_service/requirements.txt
# AI Service Dependencies

langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.10
langchain-core==0.1.10
langchain-openai==0.0.2

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

youtube-search-python==1.6.6

python-dotenv==1.0.0
pyyaml==6.0.1

pydantic==2.5.0
httpx==0.25.2

tenacity==8.2.3